--- mvc/src/test/java/samples/TestController.java
@@ -1,13 +1,0 @@
-package samples;
-
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class TestController implements Controller {
-
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
-        return "test.jsp";
-    }
-}
